Organizing code 1
