+++ internal/wink/constant.go
@@ -1,7 +1,7 @@
 package wink
 
 // minimal count that is required to be selected on a list
-const MIN_PLAYER_CNT = 2 // TODO: update to 3
+const MIN_PLAYER_CNT = 3
 
 // maximum index query count
 const MAX_MEMBER_GET int = 50
